@@ -1,69 +1,61 @@
 {
-  "name": "schedule-web-material",
-  "version": "0.0.0",
-  "license": "MIT",
-  "scripts": {
-    "ng": "ng",
-    "start": "ng serve",
-    "build": "ng build",
-    "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e"
-  },
-  "private": true,
-  "dependencies": {
-    "@angular/animations": "^5.2.5",
-    "@angular/cdk": "^5.2.1",
-    "@angular/cli": "^1.6.2",
-    "@angular/common": "^5.2.5",
-    "@angular/compiler": "^5.2.5",
-    "@angular/core": "^5.2.5",
-    "@angular/flex-layout": "^2.0.0-beta.12",
-    "@angular/forms": "^5.2.5",
-    "@angular/http": "^5.2.5",
-    "@angular/material": "^5.2.1",
-    "@angular/material-moment-adapter": "^5.2.1",
-    "@angular/platform-browser": "^5.2.5",
-    "@angular/platform-browser-dynamic": "^5.2.5",
-    "@angular/router": "^5.2.5",
-    "@swimlane/ngx-charts": "^7.0.1",
-    "@swimlane/ngx-datatable": "^11.1.4",
-    "angular-calendar": "^0.22.1",
-    "calendar-utils": "^0.0.58",
-    "core-js": "^2.5.6",
-    "coveralls": "^3.0.0",
-    "d3": "^4.12.0",
-    "date-fns": "^1.29.0",
-    "faker": "^4.1.0",
-    "firebase-tools": "^3.16.0",
-    "hammerjs": "^2.0.8",
-    "moment": "^2.20.1",
-    "ng2-completer": "^1.3.1",
-    "rxjs": "^5.5.0",
-    "ts-md5": "^1.2.4",
-    "zone.js": "^0.8.18"
-  },
-  "devDependencies": {
-    "@angular/compiler-cli": "^5.2.5",
-    "@angularclass/hmr": "^2.1.3",
-    "@types/d3": "^4.12.0",
-    "@types/jasmine": "2.8.8",
-    "@types/node": "~10.1.0",
-    "codelyzer": "~4.0.2",
-    "jasmine-core": "~2.8.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~1.7.1",
-    "karma-chrome-launcher": "^2.2.0",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.4.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.3.2",
-    "ts-node": "~3.3.0",
-    "tslint": "~5.8.0",
-    "typescript": "~2.6.2"
-  },
-  "main": "src/index.html",
-  "repository": "https://andreroggeri@gitlab.com/odonto-helper/schedule-client.git",
-  "author": "André Roggeri Campos <a.roggeri.c@gmail.com>"
+    "name": "webapp",
+    "version": "0.0.0",
+    "scripts": {
+        "ng": "ng",
+        "start": "ng serve",
+        "build": "ng build",
+        "test": "ng test",
+        "lint": "ng lint",
+        "e2e": "ng e2e"
+    },
+    "private": true,
+    "dependencies": {
+        "@angular/animations": "^6.1.4",
+        "@angular/cdk": "^6.4.6",
+        "@angular/common": "^6.1.0",
+        "@angular/compiler": "^6.1.0",
+        "@angular/core": "^6.1.0",
+        "@angular/flex-layout": "^6.0.0-beta.17",
+        "@angular/forms": "^6.1.0",
+        "@angular/http": "^6.1.0",
+        "@angular/material": "^6.4.6",
+        "@angular/material-moment-adapter": "^6.4.6",
+        "@angular/platform-browser": "^6.1.0",
+        "@angular/platform-browser-dynamic": "^6.1.0",
+        "@angular/router": "^6.1.0",
+        "@swimlane/ngx-charts": "^9.0.0",
+        "@swimlane/ngx-datatable": "^13.1.0",
+        "@types/webpack-env": "^1.13.6",
+        "angular-calendar": "^0.25.2",
+        "core-js": "^2.5.4",
+        "d3": "^5.6.0",
+        "date-fns": "^1.29.0",
+        "faker": "^4.1.0",
+        "moment": "^2.22.2",
+        "rxjs": "^6.0.0",
+        "ts-md5": "^1.2.4",
+        "zone.js": "~0.8.26"
+    },
+    "devDependencies": {
+        "@angular-devkit/build-angular": "~0.7.0",
+        "@angular/cli": "~6.1.4",
+        "@angular/compiler-cli": "^6.1.0",
+        "@angular/language-service": "^6.1.0",
+        "@types/jasmine": "~2.8.6",
+        "@types/jasminewd2": "~2.0.3",
+        "@types/node": "^10.7.1",
+        "codelyzer": "~4.2.1",
+        "jasmine-core": "~2.99.1",
+        "jasmine-spec-reporter": "~4.2.1",
+        "karma": "~1.7.1",
+        "karma-chrome-launcher": "~2.2.0",
+        "karma-coverage-istanbul-reporter": "~2.0.0",
+        "karma-jasmine": "~1.1.1",
+        "karma-jasmine-html-reporter": "^0.2.2",
+        "protractor": "~5.4.0",
+        "ts-node": "~5.0.1",
+        "tslint": "~5.9.1",
+        "typescript": "~2.7.2"
+    }
 }
