@@ -1,63 +1,55 @@
-import { ClinicService, IClinic } from 'app/clinic/clinic.service';
-import { ChangeDetectionStrategy, Component, OnInit, ViewChild } from '@angular/core';
+import { Component, OnInit, ViewChild } from '@angular/core';
 
-import { ClinicFilter } from './clinic.service';
+import { ClinicFilter, IClinic, ClinicService } from './clinic.service';
 import { Router } from '@angular/router';
-import { DataSource, CollectionViewer } from '@angular/cdk/collections';
-import { Observable } from 'rxjs/Observable';
-import { MatPaginator } from '@angular/material';
+import { MatPaginator, MatTableDataSource } from '@angular/material';
+import { Observable } from 'rxjs';
+import { startWith, switchMap, finalize, map } from 'rxjs/operators';
 
 @Component({
-    changeDetection: ChangeDetectionStrategy.OnPush,
     selector: 'app-clinic',
     templateUrl: './clinic.component.html',
     styleUrls: ['./clinic.component.scss']
 })
 export class ClinicComponent implements OnInit {
-    clinics: IClinic[] = [];
-    isLoading = false;
     @ViewChild(MatPaginator) paginator: MatPaginator;
-    dataSource: ClinicDataSource;
-    displayedColumns = ['name', 'actions']
+    isLoading = false;
+    count = 0;
+    dataSource: MatTableDataSource<IClinic>;
+    displayedColumns = ['name', 'actions'];
+    watcher: Observable<any>;
 
     constructor(private clinicService: ClinicService, private router: Router) {
     }
 
     ngOnInit() {
-        this.dataSource = new ClinicDataSource(this.clinicService, this.paginator);
+        this.dataSource = new MatTableDataSource();
+        this.watcher = this.paginator.page
+            .pipe(
+                startWith(null),
+                switchMap(() => {
+                    this.isLoading = true;
+                    const clinicFilter = new ClinicFilter();
+                    if (this.paginator.pageSize) {
+                        const offset = this.paginator.pageSize * this.paginator.pageIndex;
+                        clinicFilter.setFilterValue('pageSize', this.paginator.pageSize.toString());
+                        clinicFilter.setFilterValue('offset', offset.toString());
+                    }
+                    return this.clinicService.getAll(clinicFilter).pipe(
+                        finalize(() => this.isLoading = false),
+                        map(response => {
+                            this.count = response.count;
+                            return response.results;
+                        })
+                    );
+                }));
+        this.watcher.subscribe(response => {
+            this.dataSource.data = response;
+        });
+
     }
+
     view(clinic: IClinic) {
         this.router.navigate(['/clinicas/' + clinic.id]);
     }
-}
-
-
-class ClinicDataSource extends DataSource<IClinic> {
-    count = 0;
-    isLoading = true;
-
-    constructor(private clinicService: ClinicService, private paginator: MatPaginator) {
-        super();
-    }
-
-    connect(_collectionViewer: CollectionViewer): Observable<IClinic[]> {
-        const displayDataChages = [this.paginator.page];
-
-        return Observable.merge(...displayDataChages)
-            .startWith(null)
-            .switchMap(() => {
-                this.isLoading = true;
-                const clinicFilter = new ClinicFilter();
-                const offset = this.paginator.pageSize * this.paginator.pageIndex;
-                clinicFilter.setFilterValue('pageSize', this.paginator.pageSize.toString());
-                clinicFilter.setFilterValue('offset', offset.toString());
-                return this.clinicService.getAll(clinicFilter).map(response => {
-                    this.count = response.count;
-                    return response.results;
-                }).finally(() => this.isLoading = false);
-            })
-    }
-
-    disconnect(_collectionViewer: CollectionViewer): void { }
-
 }
