@@ -3,17 +3,17 @@ import { RouterModule, Routes } from '@angular/router';
 import { AboutComponent } from './about.component';
 import { CommonModule } from '@angular/common';
 import { FlexLayoutModule } from '@angular/flex-layout';
-import { MaterialAppModule } from 'app/shared/material.app.module';
 import { NgModule } from '@angular/core';
+import { MaterialAppModule } from '../shared/material.app.module';
 
 export const aboutRoutes: Routes = [
-  { path: 'sobre', component: AboutComponent }
+    { path: 'sobre', component: AboutComponent }
 ];
 
 @NgModule({
-  imports: [
-    CommonModule, MaterialAppModule, RouterModule, FlexLayoutModule
-  ],
-  declarations: [AboutComponent]
+    imports: [
+        CommonModule, MaterialAppModule, RouterModule, FlexLayoutModule
+    ],
+    declarations: [AboutComponent]
 })
 export class AboutModule { }
