@@ -3,8 +3,9 @@ import { Component, ViewChild, HostListener } from '@angular/core';
 import { LoginService } from './login/login.service';
 import { MatSidenav } from '@angular/material';
 import { Md5 } from 'ts-md5/dist/md5';
+import { fromEvent } from 'rxjs';
+import { map, startWith, debounceTime } from 'rxjs/operators';
 import { Router } from '@angular/router';
-import './shared/rxjs.operators';
 
 interface IMenu {
     name: string;
