@@ -1,7 +1,5 @@
-import { AboutModule, aboutRoutes } from 'app/about/about.module';
 import { AccountModule, accountRoutes } from './account/account.module';
 import { ClinicModule, clinicRoutes } from './clinic/clinic.module';
-import { DashboardModule, dashboardRoutes } from 'app/dashboard/dashboard.module';
 import { LoginModule, loginRoutes } from './login/login.module';
 import { PatientModule, patientRoutes } from './patient/patient.module';
 import { RouterModule, Routes } from '@angular/router';
@@ -12,18 +10,19 @@ import { BrowserModule } from '@angular/platform-browser';
 import { FormsModule } from '@angular/forms';
 import { MaterialAppModule } from './shared/material.app.module';
 import { NgModule, LOCALE_ID } from '@angular/core';
-import { RegisterModule } from './register/register.module';
-import { registerRoutes } from 'app/register/register.module';
+import { RegisterModule, registerRoutes } from './register/register.module';
 import { registerLocaleData } from '@angular/common';
 import localePt from '@angular/common/locales/pt';
 import localePtExtras from '@angular/common/locales/extra/pt';
 // import { MatNativeDateModule } from '@angular/material';
-import { MatMomentDateModule } from '@angular/material-moment-adapter';
 import { HttpClientModule } from '@angular/common/http';
 import { HTTP_INTERCEPTORS } from '@angular/common/http';
-import { TokenInterceptor } from 'app/shared/interceptors/token.interceptor';
-import { ErrorInterceptor } from 'app/shared/interceptors/error.interceptor';
-import { TokenService } from 'app/shared/services/token.service';
+import { dashboardRoutes, DashboardModule } from './dashboard/dashboard.module';
+import { aboutRoutes, AboutModule } from './about/about.module';
+import { MatMomentDateModule } from '@angular/material-moment-adapter';
+import { TokenService } from './shared/services/token.service';
+import { TokenInterceptor } from './shared/interceptors/token.interceptor';
+import { ErrorInterceptor } from './shared/interceptors/error.interceptor';
 
 registerLocaleData(localePt, localePtExtras);
 
