@@ -1,14 +1,14 @@
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { ComponentFixture, TestBed, async } from '@angular/core/testing';
-import { RouterLinkStubDirective, RouterStub } from 'app/shared/testing/stubs/router.stub';
 
 import { ClinicComponent } from './clinic.component';
 import { ClinicService } from './clinic.service';
 import { DataTablePagerModule } from '../shared/components/pager/datatable-pager.module';
 import { MaterialAppModule } from '../shared/material.app.module';
-import { NgxDatatableModule } from '@swimlane/ngx-datatable';
 import { Router } from '@angular/router';
-import { ClinicServiceStub } from 'app/shared/testing/stubs/clinic.stub';
+import { RouterLinkStubDirective, RouterStub } from '../shared/testing/stubs/router.stub';
+import { ClinicServiceStub } from '../shared/testing/stubs/clinic.stub';
+import { NgxDatatableModule } from '@swimlane/ngx-datatable';
 
 describe('ClinicComponent', () => {
     let component: ClinicComponent;
