@@ -1,16 +1,16 @@
-import { AuthGuard } from 'app/shared/guards/auth.guard';
 import { BrowserModule } from '@angular/platform-browser';
-import { DashboardComponent } from 'app/dashboard/dashboard.component';
 import { DirectivesModule } from '../shared/directives/directives.module';
 import { FlexLayoutModule } from '@angular/flex-layout';
 import { FormsModule } from '@angular/forms';
-import { MaterialAppModule } from 'app/shared/material.app.module';
 import { NgModule } from '@angular/core';
 import { Routes } from '@angular/router';
-import { ScheduleService } from 'app/schedule/schedule.service';
-import { SharedComponentsModule } from 'app/shared/components/shared-components.module';
 import { NgxChartsModule } from '@swimlane/ngx-charts';
-import { PatientService } from 'app/patient/patient.service';
+import { DashboardComponent } from './dashboard.component';
+import { AuthGuard } from '../shared/guards/auth.guard';
+import { MaterialAppModule } from '../shared/material.app.module';
+import { SharedComponentsModule } from '../shared/components/shared-components.module';
+import { ScheduleService } from '../schedule/schedule.service';
+import { PatientService } from '../patient/patient.service';
 
 export const dashboardRoutes: Routes = [
     { path: 'dashboard', component: DashboardComponent, canActivate: [AuthGuard] },
