@@ -1,7 +1,7 @@
-import { BaseService } from 'app/shared/services/base.service';
 import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
+import { BaseService } from '../shared/services/base.service';
+import { Observable, of } from 'rxjs';
 
 @Injectable()
 export class LoginService extends BaseService {
@@ -38,4 +38,41 @@ export class LoginService extends BaseService {
     getLocalUserInfo() {
         return JSON.parse(localStorage.getItem('user_info'));
     }
+}
+
+@Injectable()
+export class LoginServiceStub implements ILoginService {
+    authenticate(_formData: { email: string; password: string; }): Observable<any> {
+        return of([{ auth_token: 'TEST_TOKEN' }]);
+    }
+
+    isLogged(): boolean {
+        return !!localStorage.getItem('auth_token');
+    }
+
+    getUserInfo() {
+        if (localStorage.getItem('user_info')) {
+            return JSON.parse(localStorage.getItem('user_info'));
+        } else {
+            localStorage.setItem('user_info', JSON.stringify({
+                first_name: 'John',
+                last_name: 'Snow',
+                cro: '123456',
+                cro_state: 'SP',
+                sex: 'M',
+                id: 1,
+                email: 'john@snow.com'
+            }));
+            return JSON.parse(localStorage.getItem('user_info'));
+        }
+
+    }
+
+
+}
+
+interface ILoginService {
+    authenticate(formData: { email: string, password: string }): Observable<any>;
+    isLogged(): boolean;
+    getUserInfo();
 }
