@@ -3,7 +3,8 @@ import { Component, OnInit } from '@angular/core';
 import { CustomFB, CustomFG } from '../shared/validation';
 import { IMatcher, getMatchedField, getReversedMatchField } from '../shared/util';
 import { IPatient, PatientFilter, PatientService } from './patient.service';
-import { IClickEvent } from 'app/shared/components/pager/datatable-pager.component';
+import { IClickEvent } from '../shared/components/pager/datatable-pager.component';
+import { finalize } from 'rxjs/operators';
 
 @Component({
     selector: 'app-patient',
@@ -21,15 +22,15 @@ export class PatientComponent implements OnInit {
     urlFilters: IMatcher[] = [
         { prettyName: 'nome', name: 'fullName' },
         { prettyName: 'telefone', name: 'phone' },
-    ]
+    ];
 
 
     constructor(private patientService: PatientService, private router: Router, private route: ActivatedRoute) {
-        const fb = new CustomFB()
+        const fb = new CustomFB();
         this.filterForm = fb.group({
             field: ['fullName'],
             value: ['']
-        })
+        });
     }
 
     ngOnInit() {
@@ -42,12 +43,12 @@ export class PatientComponent implements OnInit {
         this.patientFilter.setFilterValue('orderBy', this.sortBy);
 
         this.patientService.getAll(this.patientFilter)
-            .finally(() => this.isLoading = false)
+            .pipe(finalize(() => this.isLoading = false))
             .subscribe(
-            response => {
-                this.patientCount = response.count;
-                this.patients = response.results;
-            });
+                response => {
+                    this.patientCount = response.count;
+                    this.patients = response.results;
+                });
     }
 
     view(selectedRow: IClickEvent<IPatient>) {
@@ -62,18 +63,18 @@ export class PatientComponent implements OnInit {
 
     sort(sortEvent: { newValue: string, column: { prop: string } }) {
         this.sortBy = sortEvent.newValue === 'asc' ? sortEvent.column.prop : '-' + sortEvent.column.prop;
-        this.getPatients(0)
+        this.getPatients(0);
     }
 
     filter() {
         const field = getReversedMatchField(this.filterForm.value.field, this.urlFilters);
-        this.router.navigate(['/pacientes', field, this.filterForm.value.value])
+        this.router.navigate(['/pacientes', field, this.filterForm.value.value]);
     }
 
     setupUrlFilterListener() {
         this.route.params.subscribe(
             params => {
-                this.patientFilter.reset()
+                this.patientFilter.reset();
                 if (params.field !== undefined && params.value !== undefined) {
                     const field = getMatchedField(params.field, this.urlFilters);
                     this.patientFilter.setFilterValue(field, params.value);
@@ -82,6 +83,6 @@ export class PatientComponent implements OnInit {
                 }
                 this.getPatients(0);
             }
-        )
+        );
     }
 }
