@@ -5,15 +5,15 @@ import { ConfirmComponent } from './confirm.component';
 import { DentistService } from '../shared/services/dentist.service';
 import { DirectivesModule } from '../shared/directives/directives.module';
 import { FlexLayoutModule } from '@angular/flex-layout';
-import { MaterialAppModule } from 'app/shared/material.app.module';
 import { NgModule } from '@angular/core';
-import { RegisterComponent } from 'app/register/register.component';
+import { RegisterComponent } from './register.component';
 import { Routes } from '@angular/router';
+import { MaterialAppModule } from '../shared/material.app.module';
 
 export const registerRoutes: Routes = [
     { path: 'cadastro', component: RegisterComponent },
     { path: 'cadastro/ativar/:uid/:token', component: ConfirmComponent },
-]
+];
 
 @NgModule({
     imports: [
