@@ -1,14 +1,14 @@
 import * as moment from 'moment';
 
 import { ActivatedRoute, Router } from '@angular/router';
-import { ActivatedRouteStub, RouterLinkStubDirective, RouterStub } from 'app/shared/testing/stubs/router.stub';
-import { CalendarModule, MOMENT } from 'angular-calendar';
 import { ComponentFixture, TestBed, async } from '@angular/core/testing';
 
 import { MaterialAppModule } from '../shared/material.app.module';
 import { ScheduleComponent } from './schedule.component';
 import { ScheduleService } from './schedule.service';
-import { ScheduleServiceStub } from 'app/shared/testing/stubs/schedule.stub';
+import { ActivatedRouteStub, RouterLinkStubDirective, RouterStub } from '../shared/testing/stubs/router.stub';
+import { CalendarModule, MOMENT } from 'angular-calendar';
+import { ScheduleServiceStub } from '../shared/testing/stubs/schedule.stub';
 
 describe('ScheduleComponent', () => {
     let component: ScheduleComponent;
@@ -41,10 +41,10 @@ describe('ScheduleComponent', () => {
     });
 
     it('should load the date from URL', () => {
-        route.testParams = { view: 'dia', date: '2017-11-13' }
+        route.testParams = { view: 'dia', date: '2017-11-13' };
         expect(component.scheduleFilter.fields.find(v => v.name === 'startDate').value).toBe('2017-11-13');
         expect(component.scheduleFilter.fields.find(v => v.name === 'endDate').value).toBe('2017-11-13');
-        route.testParams = { view: 'semana', date: '2017-11-15' }
+        route.testParams = { view: 'semana', date: '2017-11-15' };
         expect(component.scheduleFilter.fields.find(v => v.name === 'startDate').value).toBe('2017-11-12');
         expect(component.scheduleFilter.fields.find(v => v.name === 'endDate').value).toBe('2017-11-18');
     });
