@@ -1,8 +1,8 @@
 import { CommonModule } from '@angular/common';
 import { DataTablePagerComponent } from './datatable-pager.component';
 import { FlexLayoutModule } from '@angular/flex-layout';
-import { MaterialAppModule } from 'app/shared/material.app.module';
 import { NgModule } from '@angular/core';
+import { MaterialAppModule } from '../../material.app.module';
 
 @NgModule({
     imports: [CommonModule, FlexLayoutModule, MaterialAppModule],
