@@ -1,8 +1,9 @@
 import { Injectable } from '@angular/core';
 import { HttpRequest, HttpHandler, HttpEvent, HttpInterceptor, HttpResponse, HttpErrorResponse } from '@angular/common/http';
 
-import { Observable } from 'rxjs/Observable';
 import { MatSnackBar } from '@angular/material';
+import { Observable } from 'rxjs';
+import { tap } from 'rxjs/operators';
 
 
 @Injectable()
@@ -12,16 +13,16 @@ export class ErrorInterceptor implements HttpInterceptor {
 
     intercept(request: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
 
-        return next.handle(request).do((event: HttpEvent<any>) => {
+        return next.handle(request).pipe(tap((event: HttpEvent<any>) => {
             if (event instanceof HttpResponse) {
                 // Do nothing for now
             }
         }, (err: any) => {
             if (err instanceof HttpErrorResponse) {
                 if (err.status !== 400) {
-                    this.snackBar.open('Aconteceu algo errado. Tente efetuar login novamente', 'Fechar', { duration: 3000 })
+                    this.snackBar.open('Aconteceu algo errado. Tente efetuar login novamente', 'Fechar', { duration: 3000 });
                 }
             }
-        });
+        }));
     }
 }
