@@ -1,6 +1,6 @@
 import { Directive, HostListener, Injectable, Input } from '@angular/core';
+import { BehaviorSubject } from 'rxjs';
 
-import { BehaviorSubject } from 'rxjs/BehaviorSubject';
 
 @Injectable()
 export class RouterStub {
